requirements.txt
dash== 2.17.1
plotly==5.22.0
pandas==2.2.2
dash-bootstrap-components==1.6.0
dash-leaflet==1.0.15
dash-extensions==1.0.16
python-dotenv==1.0.1
dash-auth==2.3.0
